backend/requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
Django==5.2.10
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
Pillow==11.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
redis==7.1.0
sqlparse==0.5.5
